third_party/blink/web_tests/platform/generic/fast/filesystem/entry-points-missing-arguments-expected.txt

Ensures that we get exceptions when we do not include required arguments for webkitRequestFileSystem and webkitResolveLocalFileSystemURL methods.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS webkitResolveLocalFileSystemURL() threw exception TypeError: Failed to execute 'webkitResolveLocalFileSystemURL' on 'Window': 2 arguments required, but only 0 present..
PASS webkitResolveLocalFileSystemURL("") threw exception TypeError: Failed to execute 'webkitResolveLocalFileSystemURL' on 'Window': 2 arguments required, but only 1 present..
PASS webkitRequestFileSystem() threw exception TypeError: Failed to execute 'webkitRequestFileSystem' on 'Window': 3 arguments required, but only 0 present..
PASS webkitRequestFileSystem(TEMPORARY) threw exception TypeError: Failed to execute 'webkitRequestFileSystem' on 'Window': 3 arguments required, but only 1 present..
PASS webkitRequestFileSystem(TEMPORARY, 1) threw exception TypeError: Failed to execute 'webkitRequestFileSystem' on 'Window': 3 arguments required, but only 2 present..
PASS successfullyParsed is true

TEST COMPLETE
